+++ src/DTOs/category.dto.ts
@@ -1,0 +1,9 @@
+export interface CreateCategoryDto {
+  category_name: string;
+  category_description?: string;
+}
+
+export interface UpdateCategoryDto {
+  category_name?: string;
+  category_description?: string;
+}
