+++ src/DTOs/quiz-attempt.dto.ts
@@ -1,0 +1,8 @@
+export interface CreateQuizAttemptDto {
+  user_id: number;
+  quiz_id: number;
+}
+
+export interface CompleteQuizAttemptDto {
+  end_time: Date;
+}
