+++ src/DTOs/user-response.dto.ts
@@ -1,0 +1,5 @@
+export interface CreateUserResponseDto {
+  attempt_id: number;
+  question_id: number;
+  answer_id: number;
+}
