+++ src/services/difficulty.service.ts
@@ -1,4 +1,5 @@
-import { DifficultyLevelModel, DifficultyLevel, CreateDifficultyLevelDto, UpdateDifficultyLevelDto } from '../models/difficulty.model';
+import { DifficultyLevelModel, DifficultyLevel } from '../models/difficulty.model';
+import { CreateDifficultyLevelDto, UpdateDifficultyLevelDto } from '../DTOs/difficulty.dto';
 import { ErrorUtils } from '../utils/error.utils';
 
 export class DifficultyService {
